Changed center type to int.

+++ fury/shaders/interaction/visible_range.glsl
@@ -1,4 +1,4 @@
-bool inVisibleRange(vec3 center)
+bool inVisibleRange(ivec3 center)
 {
     bool xVal = lowRanges.x <= center.x && center.x <= highRanges.x;
     bool yVal = lowRanges.y <= center.y && center.y <= highRanges.y;
